requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CORS and middleware
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# AI and language models
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.8

# Google services and YouTube API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Image processing and generation
Pillow==10.1.0
requests==2.31.0

# Video processing
moviepy==1.0.3

# Azure services
azure-storage-blob==12.19.0

# Utilities
uuid4==1.1.2
asyncio-compat==0.1.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
